Add files via upload
In this project, I have built various model of the ‘Fetal Health Classification’ dataset which contains various information of the fetus. Fetal mortality is a public health issue that put in risk the women’s or baby’s life. This project uses several Machine Learning Classification techniques in order to predict the risk level of the fetal health ,it attempts to classify that into three classes namely-
Normal
Suspect
Pathological
Cardiotocography (CTG) is used during pregnancy to monitor fetal
heart rate and uterine contractions. It is monitor fetal well-being 
and allows early detection of fetal distress. CTG interpretation helps 
in determining if the pregnancy is high or low risk. 
In this project, I created models to classify the outcome of 
Cardiotocogram test to ensure the well being of the fetus.